support: Introduce AppListPreference for a much cleaner look
Something like HeadsUp Blacklist or Gaming Mode show
an app list to the user of those which have been
selected in the preference. This is where there is no
control over the sizing or layout of things. See the
following image:

BEFORE: https://imgur.com/7muwWZ7.jpg

To address this, we will simply extend the Preference
class and add a layout for it to show things properly.
See the following image:

AFTER:  https://imgur.com/DzgVOjX.jpg

Signed-off-by: calebcabob <calphonic@gmail.com>
Change-Id: I00671e0398c5b6b411d83a9e56179964445132ce